CB-373 refactor: add useConfirm file
useConfirm 훅을 다른 파일로 분리
