+++ src/components/Confirm/index.tsx
@@ -1,4 +1,4 @@
-import ConfirmModal, { useConfirm } from './ConfirmModal';
+import ConfirmModal from './ConfirmModal';
 import ConfirmPortal from './ConfirmPortal';
 
-export { ConfirmModal, useConfirm, ConfirmPortal };
+export { ConfirmModal, ConfirmPortal };
