+++ src/pages/Mypage/pets/[id].tsx
@@ -16,6 +16,9 @@ import {
 } from '@/store/api/petApi';
 import useSelectImage from '@/utils/hooks/useSelectImage';
 import useBottomSheet from '@/utils/hooks/useBottomSheet';
+import useConfirm from '@/utils/hooks/useConfirm';
+import useToastMessage from '@/utils/hooks/useToastMessage';
+import { getFileFromObjectURL } from '@/utils/libs/getFileFromObjectURL';
 import BreedBottomSheet from '@/components/BottomSheet/BreedBottomSheet';
 import BirthdayBottomSheet from '@/components/BottomSheet/BirthdayBottomSheet';
 import MonthsAgeBottomSheet from '@/components/BottomSheet/MonthsAgeBottomSheet';
@@ -27,10 +30,7 @@ import { ReactComponent as CalendarIcon } from '@/assets/icon/calendar_icon.svg'
 
 import PetDefault from '@/assets/image/pet_default.png';
 
-import { getFileFromObjectURL } from '@/utils/libs/getFileFromObjectURL';
-import useToastMessage from '@/utils/hooks/useToastMessage';
 import { useAppSelector } from '@/store/config';
-import { useConfirm } from '@/components/Confirm';
 import {
   AgeDescription,
   AgeSelectButton,
